HSEARCH-3909 Make predicates/sorts/projections/aggregation tests a *bit* more consistent
To help with testing traits (see following commits)
