+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/predicate/KnnPredicateBaseIT.java
@@ -52,7 +52,8 @@ class KnnPredicateBaseIT {
 						ScoreConfigured.index,
 						InObjectFieldConfigured.mainIndex, InObjectFieldConfigured.missingFieldIndex,
 						InvalidFieldConfigured.index, UnsupportedTypeConfigured.index,
-						SearchableConfigured.searchableYesIndex, SearchableConfigured.searchableNoIndex
+						SearchableConfigured.searchableDefaultIndex, SearchableConfigured.searchableYesIndex,
+						SearchableConfigured.searchableNoIndex
 				)
 				.setup();
 
@@ -160,7 +161,7 @@ class KnnPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:knn";
 		}
 	}
@@ -192,7 +193,7 @@ class KnnPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:knn";
 		}
 	}
@@ -203,6 +204,9 @@ class KnnPredicateBaseIT {
 	}
 
 	abstract static class SearchableConfigured extends AbstractPredicateSearchableIT {
+		private static final SimpleMappedIndex<SearchableDefaultIndexBinding> searchableDefaultIndex =
+				SimpleMappedIndex.of( root -> new SearchableDefaultIndexBinding( root, supportedFieldTypes ) )
+						.name( "searchableDefault" );
 		private static final SimpleMappedIndex<SearchableYesIndexBinding> searchableYesIndex =
 				SimpleMappedIndex.of( root -> new SearchableYesIndexBinding( root, supportedFieldTypes ) )
 						.name( "searchableYes" );
@@ -214,7 +218,7 @@ class KnnPredicateBaseIT {
 		private static final List<Arguments> parameters = new ArrayList<>();
 		static {
 			for ( FieldTypeDescriptor<?, ?> fieldType : supportedFieldTypes ) {
-				parameters.add( Arguments.of( searchableYesIndex, searchableNoIndex, fieldType ) );
+				parameters.add( Arguments.of( searchableDefaultIndex, searchableYesIndex, searchableNoIndex, fieldType ) );
 			}
 		}
 
@@ -235,7 +239,7 @@ class KnnPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:knn";
 		}
 	}
