+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/predicate/WildcardPredicateBaseIT.java
@@ -57,7 +57,8 @@ class WildcardPredicateBaseIT {
 						InObjectFieldConfigured.mainIndex, InObjectFieldConfigured.missingFieldIndex,
 						ScoreConfigured.index,
 						InvalidFieldConfigured.index, UnsupportedTypeConfigured.index,
-						SearchableConfigured.searchableYesIndex, SearchableConfigured.searchableNoIndex,
+						SearchableConfigured.searchableDefaultIndex, SearchableConfigured.searchableYesIndex,
+						SearchableConfigured.searchableNoIndex,
 						ArgumentCheckingConfigured.index,
 						TypeCheckingNoConversionConfigured.index, TypeCheckingNoConversionConfigured.compatibleIndex,
 						TypeCheckingNoConversionConfigured.rawFieldCompatibleIndex,
@@ -309,7 +310,7 @@ class WildcardPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:wildcard";
 		}
 	}
@@ -341,7 +342,7 @@ class WildcardPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:wildcard";
 		}
 	}
@@ -352,6 +353,9 @@ class WildcardPredicateBaseIT {
 	}
 
 	abstract static class SearchableConfigured extends AbstractPredicateSearchableIT {
+		private static final SimpleMappedIndex<SearchableDefaultIndexBinding> searchableDefaultIndex =
+				SimpleMappedIndex.of( root -> new SearchableDefaultIndexBinding( root, supportedFieldTypes ) )
+						.name( "searchableDefault" );
 		private static final SimpleMappedIndex<SearchableYesIndexBinding> searchableYesIndex =
 				SimpleMappedIndex.of( root -> new SearchableYesIndexBinding( root, supportedFieldTypes ) )
 						.name( "searchableYes" );
@@ -363,7 +367,7 @@ class WildcardPredicateBaseIT {
 		private static final List<Arguments> parameters = new ArrayList<>();
 		static {
 			for ( FieldTypeDescriptor<?, ?> fieldType : supportedFieldTypes ) {
-				parameters.add( Arguments.of( searchableYesIndex, searchableNoIndex, fieldType ) );
+				parameters.add( Arguments.of( searchableDefaultIndex, searchableYesIndex, searchableNoIndex, fieldType ) );
 			}
 		}
 
@@ -377,7 +381,7 @@ class WildcardPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:wildcard";
 		}
 	}
@@ -461,7 +465,7 @@ class WildcardPredicateBaseIT {
 		}
 
 		@Override
-		protected String predicateNameInErrorMessage() {
+		protected String predicateTrait() {
 			return "predicate:wildcard";
 		}
 	}
