+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/projection/AbstractProjectionInvalidFieldIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.tck.search.predicate;
+package org.hibernate.search.integrationtest.backend.tck.search.projection;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -12,25 +12,25 @@ import org.hibernate.search.engine.backend.document.IndexObjectFieldReference;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
 import org.hibernate.search.engine.backend.types.ObjectStructure;
-import org.hibernate.search.engine.search.predicate.dsl.SearchPredicateFactory;
+import org.hibernate.search.engine.search.projection.dsl.SearchProjectionFactory;
 import org.hibernate.search.util.common.SearchException;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.SimpleMappedIndex;
 
 import org.junit.jupiter.api.Test;
 
-public abstract class AbstractPredicateInvalidFieldIT {
+public abstract class AbstractProjectionInvalidFieldIT {
 
 	private final SimpleMappedIndex<IndexBinding> index;
 
-	protected AbstractPredicateInvalidFieldIT(SimpleMappedIndex<IndexBinding> index) {
+	protected AbstractProjectionInvalidFieldIT(SimpleMappedIndex<IndexBinding> index) {
 		this.index = index;
 	}
 
 	@Test
 	void use_unknownField() {
-		SearchPredicateFactory f = index.createScope().predicate();
+		SearchProjectionFactory<?, ?> f = index.createScope().projection();
 
-		assertThatThrownBy( () -> tryPredicate( f, "unknown_field" ) )
+		assertThatThrownBy( () -> tryProjection( f, "unknown_field" ) )
 				.isInstanceOf( SearchException.class )
 				.hasMessageContaining( "Unknown field" )
 				.hasMessageContaining( "'unknown_field'" );
@@ -38,29 +38,29 @@ public abstract class AbstractPredicateInvalidFieldIT {
 
 	@Test
 	void use_objectField_nested() {
-		SearchPredicateFactory f = index.createScope().predicate();
+		SearchProjectionFactory<?, ?> f = index.createScope().projection();
 
 		String fieldPath = index.binding().nested.relativeFieldName;
 
-		assertThatThrownBy( () -> tryPredicate( f, fieldPath ) )
+		assertThatThrownBy( () -> tryProjection( f, fieldPath ) )
 				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Cannot use '" + predicateTrait() + "' on field '" + fieldPath + "'" );
+				.hasMessageContaining( "Cannot use '" + projectionTrait() + "' on field '" + fieldPath + "'" );
 	}
 
 	@Test
 	void use_objectField_flattened() {
-		SearchPredicateFactory f = index.createScope().predicate();
+		SearchProjectionFactory<?, ?> f = index.createScope().projection();
 
 		String fieldPath = index.binding().flattened.relativeFieldName;
 
-		assertThatThrownBy( () -> tryPredicate( f, fieldPath ) )
+		assertThatThrownBy( () -> tryProjection( f, fieldPath ) )
 				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Cannot use '" + predicateTrait() + "' on field '" + fieldPath + "'" );
+				.hasMessageContaining( "Cannot use '" + projectionTrait() + "' on field '" + fieldPath + "'" );
 	}
 
-	protected abstract void tryPredicate(SearchPredicateFactory f, String fieldPath);
+	protected abstract void tryProjection(SearchProjectionFactory<?, ?> f, String fieldPath);
 
-	protected abstract String predicateTrait();
+	protected abstract String projectionTrait();
 
 	public static final class IndexBinding {
 		private final ObjectBinding nested;
