+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/projection/DistanceProjectionTypeIndependentIT.java
@@ -8,16 +8,12 @@ package org.hibernate.search.integrationtest.backend.tck.search.projection;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.hibernate.search.engine.backend.document.IndexObjectFieldReference;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement;
-import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
-import org.hibernate.search.engine.backend.types.ObjectStructure;
 import org.hibernate.search.engine.backend.types.Projectable;
 import org.hibernate.search.engine.spatial.GeoPoint;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.types.GeoPointFieldTypeDescriptor;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.SimpleFieldModel;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.extension.SearchSetupHelper;
-import org.hibernate.search.util.common.SearchException;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.SimpleMappedIndex;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.StubMappingScope;
 
@@ -45,21 +41,6 @@ class DistanceProjectionTypeIndependentIT {
 	@BeforeAll
 	static void setup() {
 		setupHelper.start().withIndex( index ).setup();
-	}
-
-	@Test
-	void unknownField() {
-		StubMappingScope scope = index.createScope();
-
-		assertThatThrownBy( () -> scope.projection()
-				.distance( "unknownField", SOME_POINT )
-		)
-				.isInstanceOf( SearchException.class )
-				.hasMessageContainingAll(
-						"Unknown field",
-						"unknownField",
-						index.name()
-				);
 	}
 
 	@Test
@@ -92,52 +73,14 @@ class DistanceProjectionTypeIndependentIT {
 				);
 	}
 
-	@Test
-	void objectField_nested() {
-		String fieldPath = index.binding().nestedObject.relativeFieldName;
-		StubMappingScope scope = index.createScope();
-
-		assertThatThrownBy( () -> scope.projection()
-				.distance( fieldPath, SOME_POINT ) )
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Cannot use 'projection:distance' on field '" + fieldPath + "'" );
-	}
-
-	@Test
-	void objectField_flattened() {
-		String fieldPath = index.binding().flattenedObject.relativeFieldName;
-		StubMappingScope scope = index.createScope();
-
-		assertThatThrownBy( () -> scope.projection()
-				.distance( fieldPath, SOME_POINT ) )
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Cannot use 'projection:distance' on field '" + fieldPath + "'" );
-	}
-
 	private static class IndexBinding {
 		final SimpleFieldModel<GeoPoint> geoPointField;
 
-		final ObjectMapping flattenedObject;
-		final ObjectMapping nestedObject;
 
 		IndexBinding(IndexSchemaElement root) {
 			geoPointField =
 					SimpleFieldModel.mapper( GeoPointFieldTypeDescriptor.INSTANCE, c -> c.projectable( Projectable.YES ) )
 							.map( root, "geoPoint" );
-
-			flattenedObject = new ObjectMapping( root, "flattenedObject", ObjectStructure.FLATTENED );
-			nestedObject = new ObjectMapping( root, "nestedObject", ObjectStructure.NESTED );
-		}
-	}
-
-	private static class ObjectMapping {
-		final String relativeFieldName;
-		final IndexObjectFieldReference self;
-
-		ObjectMapping(IndexSchemaElement parent, String relativeFieldName, ObjectStructure structure) {
-			this.relativeFieldName = relativeFieldName;
-			IndexSchemaObjectField objectField = parent.objectField( relativeFieldName, structure );
-			self = objectField.toReference();
 		}
 	}
