+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/projection/FieldProjectionTypeIndependentIT.java
@@ -8,14 +8,10 @@ package org.hibernate.search.integrationtest.backend.tck.search.projection;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.hibernate.search.engine.backend.document.IndexObjectFieldReference;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement;
-import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
-import org.hibernate.search.engine.backend.types.ObjectStructure;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.types.KeywordStringFieldTypeDescriptor;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.SimpleFieldModel;
 import org.hibernate.search.integrationtest.backend.tck.testsupport.util.extension.SearchSetupHelper;
-import org.hibernate.search.util.common.SearchException;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.SimpleMappedIndex;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.StubMappingScope;
 
@@ -44,21 +40,6 @@ class FieldProjectionTypeIndependentIT {
 	}
 
 	@Test
-	void unknownField() {
-		StubMappingScope scope = index.createScope();
-
-		assertThatThrownBy( () -> scope.projection()
-				.field( "unknownField", Object.class )
-		)
-				.isInstanceOf( SearchException.class )
-				.hasMessageContainingAll(
-						"Unknown field",
-						"unknownField",
-						index.name()
-				);
-	}
-
-	@Test
 	void nullClass() {
 		StubMappingScope scope = index.createScope();
 
@@ -73,53 +54,13 @@ class FieldProjectionTypeIndependentIT {
 				);
 	}
 
-	@Test
-	void objectField_nested() {
-		String fieldPath = index.binding().nestedObject.relativeFieldName;
-		StubMappingScope scope = index.createScope();
-
-		assertThatThrownBy( () -> scope.projection()
-				.field( fieldPath, Object.class )
-		)
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Cannot use 'projection:field' on field '" + fieldPath + "'" );
-	}
-
-	@Test
-	void objectField_flattened() {
-		String fieldPath = index.binding().flattenedObject.relativeFieldName;
-		StubMappingScope scope = index.createScope();
-
-		assertThatThrownBy( () -> scope.projection()
-				.field( fieldPath, Object.class )
-		)
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Cannot use 'projection:field' on field '" + fieldPath + "'" );
-	}
-
 	private static class IndexBinding {
 		final SimpleFieldModel<String> string1Field;
 
-		final ObjectMapping flattenedObject;
-		final ObjectMapping nestedObject;
 
 		IndexBinding(IndexSchemaElement root) {
 			string1Field = SimpleFieldModel.mapper( KeywordStringFieldTypeDescriptor.INSTANCE, c -> {} )
 					.map( root, "string1" );
-
-			flattenedObject = new ObjectMapping( root, "flattenedObject", ObjectStructure.FLATTENED );
-			nestedObject = new ObjectMapping( root, "nestedObject", ObjectStructure.NESTED );
-		}
-	}
-
-	private static class ObjectMapping {
-		final String relativeFieldName;
-		final IndexObjectFieldReference self;
-
-		ObjectMapping(IndexSchemaElement parent, String relativeFieldName, ObjectStructure structure) {
-			this.relativeFieldName = relativeFieldName;
-			IndexSchemaObjectField objectField = parent.objectField( relativeFieldName, structure );
-			self = objectField.toReference();
 		}
 	}
