+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/sort/SearchSortIT.java
@@ -45,7 +45,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 /**
  * Generic tests for sorts. More specific tests can be found in other classes,
- * such as {@link FieldSortBaseIT} or {@link DistanceSortBaseIT}.
+ * such as {@link FieldSortBaseIT} or {@link DistanceSortSpecificsIT}.
  */
 class SearchSortIT {
